reworking channels
