+++ src/main/java/io/github/realyusufismail/ydw/entities/Channel.java
@@ -38,83 +38,4 @@ import java.util.Optional;
 public interface Channel extends SnowFlake, GenericEntity {
     @NotNull
     ChannelType getType();
-
-    Optional<Guild> getGuild();
-
-    Optional<Integer> getPosition();
-
-    List<Overwrite> getPermissionOverwrites();
-
-    Optional<String> getName();
-
-    Optional<String> getTopic();
-
-    Optional<Boolean> isNSFW();
-
-    Optional<SnowFlake> getLastMessageId();
-
-    Optional<Integer> getBitrate();
-
-    Optional<Integer> getUserLimit();
-
-    Optional<Integer> getRateLimitPerUser();
-
-    List<User> getRecipients();
-
-    Optional<String> getIcon();
-
-    Optional<User> getOwner();
-
-    Optional<SnowFlake> getApplicationId();
-
-    Optional<SnowFlake> getParentId();
-
-    Optional<ZonedDateTime> getLastPinTimestamp();
-
-    Optional<String> getRTCRegion();
-
-    Optional<Integer> getVideoQualityMode();
-
-    Optional<Integer> getMessageCount();
-
-    Optional<Integer> getMemberCount();
-
-    Optional<ThreadMetadata> getThreadMetadata();
-
-    Optional<Member> getMember();
-
-    Optional<Integer> getDefaultAutoArchiveDuration();
-
-    Optional<Permission[]> getPermissions();
-
-    Optional<MessageFlags[]> getFlags();
-
-    Optional<Category> getCategory();
-
-
-    // Rest Actions
-    MessageActionReg sendMessage(String message);
-
-    MessageActionReg sendEmbedMessage(EmbedBuilder embedBuilder);
-
-    @NotNull
-    Message getMessage(long messageId);
-
-    @NotNull
-    default Message getMessage(@NotNull String messageId) {
-        return getMessage(Long.parseUnsignedLong(messageId));
-    }
-
-    List<Message> getMessages(int limit);
-
-    @NotNull
-    default MessageActionReg deleteMessage(@NotNull String messageId) {
-        return deleteMessage(Long.parseUnsignedLong(messageId));
-    }
-
-    @NotNull
-    MessageActionReg deleteMessage(long messageId);
-
-    @NotNull
-    MessageActionReg deleteMessages(int amount);
 }
